+++ R/hv.R
@@ -54,14 +54,15 @@
 #' speed.
 #'
 #' For 5D or higher, it uses a recursive algorithm
-#' \citep{FonPaqLop06:hypervolume} with \eqn{\text{HV4D}^{+}} as the base case.
-#' The original proposal \citep{FonPaqLop06:hypervolume} had the HV3D algorithm
-#' as the base case.  This recursive algorithm has \eqn{O(n^{m-2} \log n)} time
-#' and linear space complexity in the worst-case, where \eqn{m} is the
-#' dimension of the points, but experimental results show that the pruning
-#' techniques used may reduce the time complexity even further.  Andreia
-#' P. Guerreiro enhanced the numerical stability of the algorithm by avoiding
-#' floating-point comparisons of partial hypervolumes.
+#' \citep{FonPaqLop06:hypervolume} with \eqn{\text{HV4D}^{+}} as the base case,
+#' resulting in a \eqn{O(n^{d-2})} time complexity and \eqn{O(n)} space
+#' complexity in the worst-case, where \eqn{d} is the dimension of the points.
+#' Experimental results show that the pruning techniques used may reduce the
+#' time complexity even further.  The original proposal
+#' \citep{FonPaqLop06:hypervolume} had the HV3D algorithm as the base case,
+#' giving a time complexity of \eqn{O(n^{d-2} \log n)}.  Andreia P. Guerreiro
+#' enhanced the numerical stability of the algorithm by avoiding floating-point
+#' comparisons of partial hypervolumes.
 #'
 #'
 #' @references
@@ -82,7 +83,8 @@ hypervolume <- function(x, reference, maximise = FALSE)
 {
   x <- as_double_matrix(x)
   nobjs <- ncol(x)
-  if (is.null(reference)) stop("reference cannot be NULL")
+  if (!is.numeric(reference))
+    stop("a numerical reference vector must be provided")
   if (length(reference) == 1L) reference <- rep_len(reference, nobjs)
 
   if (any(maximise)) {
@@ -100,11 +102,20 @@ hypervolume <- function(x, reference, maximise = FALSE)
 
 #' Hypervolume contribution of a set of points
 #'
-#' Computes the hypervolume contribution of each point given a set of points
-#' with respect to a given reference point assuming minimization of all
-#' objectives.  Dominated points have zero contribution. Duplicated points have
-#' zero contribution even if not dominated, because removing one of them does
-#' not change the hypervolume dominated by the remaining set.
+#' Computes the hypervolume contribution of each point of a set of points with
+#' respect to a given reference point.  The hypervolume contribution of point
+#' \eqn{\vec{p} \in X} is \eqn{\text{hvc}(\vec{p}) = \text{hyp}(X) -
+#' \text{hyp}(X \setminus \{\vec{p}\})}. Dominated points have zero
+#' contribution but they may influence the contribution of other
+#' points. Duplicated points have zero contribution even if not dominated,
+#' because removing one of the duplicates does not change the hypervolume of
+#' the remaining set.
+#'
+#' The current implementation uses the \eqn{O(n\log n)} dimension-sweep
+#' algorithm for 2D and the naive algorithm that requires calculating the
+#' hypervolume \eqn{|X|+1} times for dimensions larger than 2.
+#'
+#' For details about the hypervolume, see [hypervolume()].
 #'
 #' @inheritParams hypervolume
 #'
@@ -121,6 +132,14 @@ hypervolume <- function(x, reference, maximise = FALSE)
 #' \insertRef{BeuFonLopPaqVah09:tec}{moocore}
 #'
 #' @examples
+#'
+#' x <- matrix(c(5,1, 1,5, 4,2, 4,4, 5,1), ncol=2, byrow=TRUE)
+#' hv_contributions(x, reference=c(6,6))
+#' # hvc[(5,1)] = 0 = duplicated
+#' # hvc[(1,5)] = 3 = (4 - 1) * (6 - 5)
+#' # hvc[(4,2)] = 2 = (5 - 4) * (4 - 2)
+#' # hvc[(4,4)] = 0 = dominated
+#' # hvc[(5,1)] = 0 = duplicated
 #'
 #' data(SPEA2minstoptimeRichmond)
 #' # The second objective must be maximized