+++ src/libmoocore/hv.c
@@ -4,28 +4,14 @@
 
  ---------------------------------------------------------------------
 
-                       Copyright (c) 2010
-                  Carlos M. Fonseca <cmfonsec@dei.uc.pt>
-             Manuel Lopez-Ibanez <manuel.lopez-ibanez@manchester.ac.uk>
-                    Luis Paquete <paquete@dei.uc.pt>
+                           Copyright (C) 2010, 2025
+                    Carlos M. Fonseca <cmfonsec@dei.uc.pt>
+          Manuel Lopez-Ibanez <manuel.lopez-ibanez@manchester.ac.uk>
+                       Luis Paquete <paquete@dei.uc.pt>
 
- This program is free software (software libre); you can redistribute
- it and/or modify it under the terms of the GNU General Public License
- as published by the Free Software Foundation; either version 2 of the
- License, or (at your option) any later version.
-
- This program is distributed in the hope that it will be useful, but
- WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program; if not, you can obtain a copy of the GNU
- General Public License at:
-                 http://www.gnu.org/copyleft/gpl.html
- or by writing to:
-           Free Software Foundation, Inc., 59 Temple Place,
-                 Suite 330, Boston, MA 02111-1307 USA
+ This Source Code Form is subject to the terms of the Mozilla Public
+ License, v. 2.0. If a copy of the MPL was not distributed with this
+ file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
  ----------------------------------------------------------------------
 
@@ -78,25 +64,25 @@ static int compare_node(const void *p1, const void* p2)
 
 static fpli_dlnode_t *
 fpli_setup_cdllist(const double * restrict data, dimension_t d,
-                   int * restrict size, const double * restrict ref)
+                   size_t * restrict size, const double * restrict ref)
 {
     ASSUME(d > STOP_DIMENSION);
     dimension_t d_stop = d - STOP_DIMENSION;
-    int n = *size;
-    fpli_dlnode_t *head  = malloc ((n+1) * sizeof(fpli_dlnode_t));
+    size_t n = *size;
+    fpli_dlnode_t *head = malloc ((n+1) * sizeof(*head));
+    head->next = malloc(2 * d_stop * (n+1) * sizeof(fpli_dlnode_t*));
+    head->prev = head->next + d_stop * (n+1);
+    head->area = malloc(2 * d_stop * (n+1) * sizeof(double));
+    head->vol = head->area + d_stop * (n+1);
     head->x = NULL; /* head contains no data */
     head->ignore = 0;  /* should never get used */
-    head->next = malloc(d_stop * (n+1) * sizeof(fpli_dlnode_t*));
-    head->prev = malloc(d_stop * (n+1) * sizeof(fpli_dlnode_t*));
-    head->area = malloc(d_stop * (n+1) * sizeof(double));
-    head->vol = malloc(d_stop * (n+1) * sizeof(double));
 
-    int i, j;
-    for (i = 1, j = 0; j < n; j++) {
+    size_t i = 1;
+    for (size_t j = 0; j < n; j++) {
         /* Filters those points that do not strictly dominate the reference
            point.  This is needed to assure that the points left are only those
            that are needed to calculate the hypervolume. */
-        if (unlikely(strongly_dominates(data + j * d, ref, d))) {
+        if (likely(strongly_dominates(data + j * d, ref, d))) {
             head[i].x = data + (j+1) * d; /* this will be fixed a few lines below... */
             head[i].ignore = 0;
             head[i].next = head->next + i * d_stop;
@@ -133,6 +119,7 @@ fpli_setup_cdllist(const double * restrict data, dimension_t d,
 
     free(scratch);
 
+    // FIXME: This should not be necessary.
     for (i = 0; i < d_stop; i++)
         head->area[i] = 0;
 
@@ -144,14 +131,12 @@ finish:
 static void fpli_free_cdllist(fpli_dlnode_t * head)
 {
     free(head->next);
-    free(head->prev);
     free(head->area);
-    free(head->vol);
     free(head);
 }
 
-static void delete(fpli_dlnode_t * restrict nodep, dimension_t dim,
-                   double * restrict bound)
+static void
+delete(fpli_dlnode_t * restrict nodep, dimension_t dim, double * restrict bound)
 {
     ASSUME(dim > STOP_DIMENSION);
     for (dimension_t i = STOP_DIMENSION; i < dim; i++) {
@@ -160,11 +145,11 @@ static void delete(fpli_dlnode_t * restrict nodep, dimension_t dim,
         nodep->next[d]->prev[d] = nodep->prev[d];
         if (bound[d] > nodep->x[i])
             bound[d] = nodep->x[i];
-  }
+    }
 }
 
-static void reinsert (fpli_dlnode_t * restrict nodep, dimension_t dim,
-                      double * restrict bound)
+static void
+reinsert(fpli_dlnode_t * restrict nodep, dimension_t dim, double * restrict bound)
 {
     ASSUME(dim > STOP_DIMENSION);
     for (dimension_t i = STOP_DIMENSION; i < dim; i++) {
@@ -176,7 +161,7 @@ static void reinsert (fpli_dlnode_t * restrict nodep, dimension_t dim,
     }
 }
 
-static dlnode_t *
+static void
 fpli_hv4d_setup_cdllist(const fpli_dlnode_t * restrict pp,
                         dlnode_t * restrict list, size_t n _attr_maybe_unused)
 {
@@ -210,21 +195,24 @@ fpli_hv4d_setup_cdllist(const fpli_dlnode_t * restrict pp,
     // q = last point, q->next = s3, s3->prev = last point
     q->next[d] = list+2;
     (list+2)->prev[d] = q;
-    return list;
+}
+
+static double
+one_point_hv(const double * restrict x, const double * restrict ref, dimension_t d)
+{
+    double hv = ref[0] - x[0];
+    for (dimension_t i = 1; i < d; i++)
+        hv *= (ref[i] - x[i]);
+    return hv;
 }
 
 double hv4dplusU(dlnode_t * list);
 
 static double
-fpli_hv4d(fpli_dlnode_t *list, dlnode_t * restrict list4d,
-          size_t c, const double * restrict ref)
+fpli_hv4d(fpli_dlnode_t *list, dlnode_t * restrict list4d, size_t c)
 {
-    ASSUME(c >= 1);
-    fpli_dlnode_t *pp = list->next[0];
-    if (c == 1) {
-        return (ref[0] - pp->x[0]) * (ref[1] - pp->x[1]) * (ref[2] - pp->x[2]) * (ref[3] - pp->x[3]);
-    }
-    fpli_hv4d_setup_cdllist(pp, list4d, c);
+    ASSUME(c > 1);
+    fpli_hv4d_setup_cdllist(list->next[0], list4d, c);
     double hv = hv4dplusU(list4d);
     return hv;
 }
@@ -234,129 +222,120 @@ hv_recursive(fpli_dlnode_t * restrict list, dlnode_t * restrict list4d,
              dimension_t dim, size_t c,
              const double * restrict ref, double * restrict bound)
 {
+    ASSUME(dim > STOP_DIMENSION);
+    ASSUME(c > 1);
     /* ------------------------------------------------------
        General case for dimensions higher than 4D
        ------------------------------------------------------ */
-    if ( dim > STOP_DIMENSION ) {
-        const dimension_t d_stop = dim - STOP_DIMENSION;
-        fpli_dlnode_t *p1 = list->prev[d_stop];
-        for (fpli_dlnode_t *pp = p1; pp->x; pp = pp->prev[d_stop]) {
-            if (pp->ignore < dim)
-                pp->ignore = 0;
-        }
-        fpli_dlnode_t *p0 = list;
-        while (c > 1
-               /* We delete all points x[dim] > bound[d_stop]. In case of
-                  repeated coordinates, we also delete all points
-                  x[dim] == bound[d_stop] except one. */
-               && (p1->x[dim] > bound[d_stop]
-                   || p1->prev[d_stop]->x[dim] >= bound[d_stop])
-            ) {
-            // FIXME: Instead of deleting each point, unlink the start and end
-            // nodes after the loop.
-            delete(p1, dim, bound);
-            p0 = p1;
-            p1 = p1->prev[d_stop];
-            c--;
-        }
+    const dimension_t d_stop = dim - STOP_DIMENSION;
+    fpli_dlnode_t *p1 = list->prev[d_stop];
+    for (fpli_dlnode_t *pp = p1; pp->x; pp = pp->prev[d_stop]) {
+        if (pp->ignore < dim)
+            pp->ignore = 0;
+    }
+    fpli_dlnode_t *p0 = list;
+    /* Delete all points x[dim] > bound[d_stop].  In case of repeated
+       coordinates, delete also all points x[dim] == bound[d_stop] except
+       one.  */
+    while (p1->x[dim] > bound[d_stop] || p1->prev[d_stop]->x[dim] >= bound[d_stop]) {
+        // FIXME: Instead of deleting each point, unlink the start and end
+        // nodes after the loop.
+        delete(p1, dim, bound);
+        p0 = p1;
+        p1 = p1->prev[d_stop];
+        c--;
+        if (c == 1)
+            break;
+    }
 
-        double hyperv = 0;
-        if (c > 1) {
-            hyperv = p1->prev[d_stop]->vol[d_stop] + p1->prev[d_stop]->area[d_stop]
-                * (p1->x[dim] - p1->prev[d_stop]->x[dim]);
+    double hyperv = 0;
+    if (c > 1) {
+        hyperv = p1->prev[d_stop]->vol[d_stop] + p1->prev[d_stop]->area[d_stop]
+            * (p1->x[dim] - p1->prev[d_stop]->x[dim]);
+    } else {
+        ASSUME(c == 1);
+        p1->area[0] = one_point_hv(p1->x, ref, STOP_DIMENSION);
+        for (dimension_t i = STOP_DIMENSION; i < dim; i++)
+            p1->area[i + 1 - STOP_DIMENSION] = p1->area[i - STOP_DIMENSION] * (ref[i] - p1->x[i]);
+        p1->vol[d_stop] = 0;
+        if (p0->x == NULL) {
+            return p1->area[d_stop] * (ref[dim] - p1->x[dim]);
+        }
+        hyperv = p1->area[d_stop] * (p0->x[dim] - p1->x[dim]);
+        bound[d_stop] = p0->x[dim];
+        reinsert(p0, dim, bound);
+        c++;
+        p1 = p0;
+        p0 = p0->next[d_stop];
+    }
+
+    assert(c > 1);
+    while (true) {
+        p1->vol[d_stop] = hyperv;
+        double hypera;
+        if (p1->ignore >= dim) {
+            hypera = p1->prev[d_stop]->area[d_stop];
         } else {
-            ASSUME(c == 1);
-            double area = (ref[0] - p1->x[0]);
-            for (dimension_t i = 1; i <= STOP_DIMENSION; i++)
-                area = area * (ref[i] - p1->x[i]);
-            p1->area[0] = area;
-            for (dimension_t i = 1; i <= d_stop; i++)
-                p1->area[i] = p1->area[i-1] * (ref[STOP_DIMENSION + i] - p1->x[STOP_DIMENSION + i]);
-        }
-
-        while(true) {
-            p1->vol[d_stop] = hyperv;
-            if (p1->ignore >= dim) {
-                p1->area[d_stop] = p1->prev[d_stop]->area[d_stop];
+            if (dim - 1 == STOP_DIMENSION) {
+                /*---------------------------------------
+                  base case of dimension 4
+                  --------------------------------------*/
+                hypera = fpli_hv4d(list, list4d, c);
             } else {
-                p1->area[d_stop] = hv_recursive(list, list4d, dim-1, c, ref, bound);
-                if (p1->area[d_stop] <= p1->prev[d_stop]->area[d_stop])
-                    p1->ignore = dim;
+                hypera = hv_recursive(list, list4d, dim-1, c, ref, bound);
             }
-
-            if (p0->x == NULL) {
-                hyperv += p1->area[d_stop] * (ref[dim] - p1->x[dim]);
-                return hyperv;
-            }
-            hyperv += p1->area[d_stop] * (p0->x[dim] - p1->x[dim]);
-            bound[d_stop] = p0->x[dim];
-            reinsert(p0, dim, bound);
-            c++;
-            p1 = p0;
-            p0 = p0->next[d_stop];
+            if (hypera <= p1->prev[d_stop]->area[d_stop])
+                p1->ignore = dim;
         }
+        p1->area[d_stop] = hypera;
+        if (p0->x == NULL) {
+            hyperv += hypera * (ref[dim] - p1->x[dim]);
+            return hyperv;
+        }
+        hyperv += hypera * (p0->x[dim] - p1->x[dim]);
+        bound[d_stop] = p0->x[dim];
+        reinsert(p0, dim, bound);
+        c++;
+        p1 = p0;
+        p0 = p0->next[d_stop];
     }
-
-    /* ---------------------------
-       special case of dimension 4
-       --------------------------- */
-    else if (dim == STOP_DIMENSION) {
-        return fpli_hv4d(list, list4d, c, ref);
-    }
-    else
-        fatal_error("%s:%d: unreachable condition! \n"
-                    "This is a bug, please report it to "
-                    "manuel.lopez-ibanez@manchester.ac.uk\n", __FILE__, __LINE__);
 }
 
 static double
 hv2d(const double * restrict data, size_t n, const double * restrict ref)
 {
-    const double **p = malloc (n * sizeof(*p));
+    const double **p = generate_sorted_doublep_2d(data, &n, ref[0]);
+    if (unlikely(n == 0)) return 0;
     if (unlikely(!p)) return -1;
-
-    for (size_t k = 0; k < n; k++)
-        p[k] = data + 2 * k;
-
-    qsort(p, n, sizeof(*p), &cmp_doublep_x_asc_y_asc);
 
     double hyperv = 0;
     double prev_j = ref[1];
     size_t j = 0;
     do {
         /* Filter everything that may be above the ref point. */
-        while (j < n && p[j][1] >= prev_j)
-            j++;
-        if (unlikely(j == n || p[j][0] >= ref[0]))
-            break; /* No other point dominates ref. */
-        // We found one point that dominates ref.
-        hyperv += (ref[0] - p[j][0]) * (prev_j - p[j][1]);
-        prev_j = p[j][1];
+        if (p[j][1] < prev_j) {
+            // We found one point that dominates ref.
+            hyperv += (ref[0] - p[j][0]) * (prev_j - p[j][1]);
+            prev_j = p[j][1];
+        }
         j++;
-    } while (j < n && p[j][0] < ref[0]);
+    } while (j < n);
 
     free(p);
     return hyperv;
 }
 
 double hv3d_plus(const double * restrict data, size_t n, const double * restrict ref);
-
-static double
-hv4d(const double * restrict data, size_t n, const double * restrict ref)
-{
-    dlnode_t * list = setup_cdllist(data, n, ref);
-    double hv = hv4dplusU(list);
-    free_cdllist(list);
-    return hv;
-}
+double hv4d(const double * restrict data, size_t n, const double * restrict ref);
 
 /*
    Returns 0 if no point strictly dominates ref.
    Returns -1 if out of memory.
 */
-double fpli_hv(const double * restrict data, int d, int n,
+double fpli_hv(const double * restrict data, int d, int npoints,
                const double * restrict ref)
 {
+    size_t n = (size_t) npoints;
     if (unlikely(n == 0)) return 0.0;
     ASSUME(d < 256);
     ASSUME(d > 1);
@@ -370,20 +349,19 @@ double fpli_hv(const double * restrict data, int d, int n,
         /* Returning here would leak memory.  */
         hyperv = 0.0;
     } else if (unlikely(n == 1)) {
-        const double * x = list->next[0]->x;
-        hyperv = 1;
-        for (dimension_t i = 0; i < dim; i++)
-            hyperv *= ref[i] - x[i];
+        hyperv = one_point_hv(list->next[0]->x, ref, dim);
     } else {
-        dlnode_t * list4d = new_cdllist(n, ref);
-        double * bound = malloc ( (dim - STOP_DIMENSION) * sizeof(double));
-        for (dimension_t i = 0; i < (dim - STOP_DIMENSION); i++)
+        const dimension_t d_stop = dim - STOP_DIMENSION;
+        ASSUME(d_stop > 1 && d_stop < 255); // Silence -Walloc-size-larger-than= warning
+        double * bound = malloc(d_stop * sizeof(double));
+        for (dimension_t i = 0; i < d_stop; i++)
             bound[i] = -DBL_MAX;
+        dlnode_t * list4d = new_cdllist(n, ref);
         hyperv = hv_recursive(list, list4d, dim - 1, n, ref, bound);
-        free (bound);
         free_cdllist(list4d);
+        free(bound);
     }
     /* Clean up.  */
-    fpli_free_cdllist (list);
+    fpli_free_cdllist(list);
     return hyperv;
 }