+++ src/libmoocore/hv.h
@@ -4,28 +4,14 @@
 
  ---------------------------------------------------------------------
 
-                       Copyright (c) 2005, 2006
-                  Carlos M. Fonseca <cmfonsec@dei.uc.pt>
-             Manuel Lopez-Ibanez <manuel.lopez-ibanez@manchester.ac.uk>
-                    Luis Paquete <paquete@dei.uc.pt>
+                          Copyright (C) 2005-2006, 2025
+                    Carlos M. Fonseca <cmfonsec@dei.uc.pt>
+          Manuel Lopez-Ibanez <manuel.lopez-ibanez@manchester.ac.uk>
+                       Luis Paquete <paquete@dei.uc.pt>
 
- This program is free software (software libre); you can redistribute
- it and/or modify it under the terms of the GNU General Public License
- as published by the Free Software Foundation; either version 2 of the
- License, or (at your option) any later version.
-
- This program is distributed in the hope that it will be useful, but
- WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program; if not, you can obtain a copy of the GNU
- General Public License at:
-                 http://www.gnu.org/copyleft/gpl.html
- or by writing to:
-           Free Software Foundation, Inc., 59 Temple Place,
-                 Suite 330, Boston, MA 02111-1307 USA
+ This Source Code Form is subject to the terms of the Mozilla Public
+ License, v. 2.0. If a copy of the MPL was not distributed with this
+ file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
  ----------------------------------------------------------------------
 
@@ -34,14 +20,14 @@
 #ifndef HV_H_
 #define HV_H_
 
-#ifdef __cplusplus
-extern "C" {
-#endif
+#include "libmoocore-config.h"
 
-double fpli_hv(const double *data, int d, int n, const double *ref);
-void hv_contributions (double *hvc, double *points, int dim, int size, const double * ref);
-#ifdef __cplusplus
-}
-#endif
+// C++ needs to know that types and declarations are C, not C++.
+BEGIN_C_DECLS
 
-#endif
+MOOCORE_API double fpli_hv(const double *data, int d, int n, const double *ref);
+MOOCORE_API double hv_contributions(double *hvc, double *points, int dim, int size, const double * ref);
+
+END_C_DECLS
+
+#endif // HV_H_